Separated error handling from parsing code
