+++ include/OE/Carbon/parser.h
@@ -152,90 +152,6 @@ namespace csl {
         }
     }
 
-    // Sadly, ParseError cannon inherit from std::exception, since that would not allow the
-    // return type to be a string, which is necessary in this case in order to get a proper
-    // error message (Returning "result_ss.str().c_str()" doesn't work, since it returns a
-    // pointer to an object that gets destroyed a soon as the scope of "what()" is left,
-    // thereby basically returning garbage)
-    class parser_error_t {
-    public:
-        virtual std::string what() const throw() = 0;
-        virtual ~parser_error_t()                = default;
-
-        std::string name_;
-    };
-
-    class unexpected_symbol_error_t : public parser_error_t {
-    public:
-        unexpected_symbol_error_t(std::string unexpected, std::string expected, std::size_t line_num, std::size_t col_num)
-            : unexpected_(unexpected), expected_(expected), line_num_(line_num), col_num_(col_num) {
-            name_ = "csl::unexpected_symbol_error_t";
-        };
-
-
-        std::string what() const throw() {
-            std::stringstream result_ss;
-
-            result_ss << "Unexpetcted Symbol \"" << unexpected_ << "\" ";
-            result_ss << "at line " << line_num_ << ':' << col_num_ << ": ";
-            result_ss << "Expected " << expected_;
-
-            return result_ss.str();
-        }
-
-    private:
-        std::string unexpected_;
-        std::string expected_;
-
-        const std::size_t line_num_;
-        const std::size_t col_num_;
-    };
-
-    // TODO: Put this in a more logical place
-    class interpreter_error_t {
-    public:
-        virtual std::string what() const throw() = 0;
-        virtual ~interpreter_error_t()           = default;
-        std::string name_;
-    };
-
-    class unset_object_error_t : public interpreter_error_t {
-    public:
-        unset_object_error_t(std::string_view object) : object_(object) {
-            name_ = "csl::unset_object_error_t";
-        }
-
-
-        std::string what() const throw() {
-            std::stringstream result_ss;
-
-            result_ss << "Object \"" << object_ << "\" is not contained in the";
-            result_ss << " syntax tree obtained by the parser";
-
-            return result_ss.str();
-        }
-
-    private:
-        std::string_view object_;
-    };
-
-    // TODO: Somehow show line and col number
-    // Choose ONE of the two to inherit from... this breaks error handling otherwise
-    // class semantic_error_t : parser_error_t, interpreter_error_t {
-    class semantic_error_t : public parser_error_t {
-    public:
-        semantic_error_t(const char* msg) : msg_(msg) {
-            name_ = "csl::semantic_error_t";
-        }
-
-        std::string what() const throw() {
-            return msg_;
-        }
-
-    private:
-        const char* msg_;
-    };
-
 
     /*
      *
@@ -255,7 +171,7 @@ namespace csl {
             try {
                 return map_.at(t);
             } catch (std::out_of_range& e) {
-                throw unset_object_error_t(t);
+                throw unset_object_error_t{t};
             }
         }
 
@@ -356,7 +272,6 @@ namespace csl {
 
                     // TODO: Is access by index more efficient? This option weirdly seems to
                     // not be documented as well
-                    // TODO: Is emplace_back (or the equivalent for maps) applicable here?
                     if (std::holds_alternative<single_assignment_t>(as_generic)) {
                         result.single_assignments[as_name] = std::get<single_assignment_t>(as_generic);
                     }
@@ -384,8 +299,7 @@ namespace csl {
             expect_token(token_type::identifier);
 
             if (token_it_->content != el_name)
-                throw semantic_error_t("Closing tag identifier does not match opening tag"
-                                       " identifier");
+                throw semantic_error_t{"Closing tag identifier does not match opening tag identifier"};
 
             ++token_it_;
 
@@ -396,7 +310,6 @@ namespace csl {
         }
 
         generic_assignment_t parse_assignment() {
-            // TODO: Check where to put this in order to get better performance
             expect_token(token_type::integer, token_type::floating_point, token_type::identifier, token_type::open_brace,
                          token_type::string);
 
@@ -448,10 +361,8 @@ namespace csl {
                 expected.erase(expected.size() - 2, 2);
                 std::string unexpected(token_it_->content);
 
-                // TODO
-                throw unexpected_symbol_error_t(unexpected, expected, 0, 0);
-                //                throw unexpected_symbol_error_t(unexpected, expected, lexer_.get_line_num(),
-                //                                                lexer_.get_col_num() - token_it_->content.size());
+                // TODO: Set line and column numbers
+                throw unexpected_token_error_t{unexpected, expected, 0, 0};
             }
         }
 
@@ -469,6 +380,7 @@ namespace csl {
         }
     };
 
+
     // TODO: Forwarding reference?
     inline element parse(std::string& input) {
         lexer_t lexer(input);
@@ -480,6 +392,8 @@ namespace csl {
         parser_t parser(token_range.begin(), token_range.end());
         return parser.parse();
     }
+
+
 } // namespace csl
